18: Refactoring
