--- src/main/java/com/blackducksoftware/integration/hub/connection/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Classes that manage the connection to the Hub
- * 
- * @author sbillings
- *
- */
-package com.blackducksoftware.integration.hub.connection;
