+++ src/main/java/com/blackducksoftware/integration/hub/rest/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * REST connection management /use
+ * 
+ * @author sbillings
+ *
+ */
+package com.blackducksoftware.integration.hub.rest;
